+++ src/Contracts/HasAddresses.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Creasi\Nusa\Contracts;
 
 /**
- * @property-read \Illuminate\Database\Eloquent\Collection<int, Address>|Address[] $addresses
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, Address> $addresses
+ *
+ * @mixin \Illuminate\Database\Eloquent\Model
  */
-interface Addressable
+interface HasAddresses
 {
     /**
      * @return \Illuminate\Database\Eloquent\Relations\MorphMany|Address
