+++ src/Models/Concerns/WithAddress.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Creasi\Nusa\Support;
+namespace Creasi\Nusa\Models\Concerns;
 
 /**
- * @mixin \Illuminate\Database\Eloquent\Model
+ * @mixin \Creasi\Nusa\Contracts\HasAddress
  */
-trait HasAddress
+trait WithAddress
 {
     /**
      * @return \Illuminate\Database\Eloquent\Relations\MorphOne|\Creasi\Nusa\Contracts\Address
